Add files via upload
This project extracts the in-text references from patents. We approached the problem as a sequence labeling task. Given the tokens of a patent's text, we label each token as `O' (outside of reference), `I' (inside of reference), or `B' (beginning token of a reference). 
We tested various pre-trained models based on BERT for sequence labling.